model/requirements.txt
pandas
psycopg2-binary
scikit-learn
fuzzywuzzy
python-Levenshtein
joblib
